[#113076247] Refactor docs update
Make the doc update trigger a custom event that only the DocEdit
component listens to
