+++ server/controllers/documents.js
@@ -120,6 +120,7 @@
           if (err) {
             return next(err);
           } else {
+            console.log(doc);
             // If the user is the doc owner, allow access
             if (user._id == doc.ownerId) {
               next();
